test: modified tests, renamed package
* renamed package from transactions to transaction
