+++ transaction/transaction_test.go
@@ -1,4 +1,4 @@
-package transactions
+package transaction
 
 import (
 	"testing"
